+++ src/code/Business_logic/ATM.java
@@ -1,5 +1,9 @@
+package code.Business_logic;
 // ATM.java
 // Represents an automated teller machine
+
+import code.GUI.*;
+import code.Database.*;
 
 public class ATM 
 {
