+++ src/code/Business_logic/Screen.java
@@ -1,15 +1,27 @@
-// DepositSlot.java
-// Represents the deposit slot of the ATM
+package code.Business_logic;
+// Screen.java
+// Represents the screen of the ATM
 
-public class DepositSlot 
+public class Screen
 {
-   // indicates whether envelope was received (always returns true, 
-   // because this is only a software simulation of a real deposit slot)
-   public boolean isEnvelopeReceived()
+   // displays a message without a carriage return
+   public void displayMessage( String message ) 
    {
-      return true; // deposit envelope was received
-   } // end method isEnvelopeReceived
-} // end class DepositSlot
+      System.out.print( message ); 
+   } // end method displayMessage
+
+   // display a message with a carriage return
+   public void displayMessageLine( String message ) 
+   {
+      System.out.println( message );   
+   } // end method displayMessageLine
+
+   // display a dollar amount
+   public void displayDollarAmount( double amount )
+   {
+      System.out.printf( "$%,.2f", amount );   
+   } // end method displayDollarAmount 
+} // end class Screen
