+++ src/code/Database/BankDatabase.java
@@ -1,5 +1,6 @@
 // BankDatabase.java
 // Represents the bank account information database 
+package code.Database;
 
 public class BankDatabase
 {
