+++ src/code/GUI/BalanceInquiry.java
@@ -1,5 +1,9 @@
+package code.GUI;
 // BalanceInquiry.java
 // Represents a balance inquiry ATM transaction
+
+import code.Business_logic.*;
+import code.Database.*;
 
 public class BalanceInquiry extends Transaction
 {
