+++ src/code/GUI/CashDispenser.java
@@ -1,3 +1,4 @@
+package code.GUI;
 // CashDispenser.java
 // Represents the cash dispenser of the ATM
